% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateGCContent.R
\name{calculateGCContent}
\alias{calculateGCContent}
\title{Calculate GC Content of Extracted Sequences}
\usage{
calculateGCContent(sequences)
}
\arguments{
\item{sequences}{A character vector of DNA sequences, typically obtained from the \code{extractMutationSequences} function.}
}
\value{
A numeric vector with GC content percentages for each sequence in the input.
}
\description{
This function calculates the GC content (the percentage of G and C bases) for each sequence around a mutation site
with help of the Biostrings package.
}
\details{
This function assumes that the input sequences are in a character vector format, with each element representing a DNA sequence.
The GC content is calculated as the percentage of 'G' and 'C' bases in each sequence, relative to the total length of the sequence.
The function relies on the Biostrings package for efficient sequence handling and GC content calculation.
}
\examples{
\dontrun{
# Example 1
# Calculate GC content for a predefined set of sequences
sequences <- c("ATGCGT", "CGCGGC", "ATATAT")
gc_content <- calculateGCContent(sequences)
print(gc_content)

# Example 2
# Extract sequences around mutation sites and calculate their GC content
library(BSgenome.Hsapiens.UCSC.hg19)
library(GenomicRanges)

# Extract sequences
extracted_data <- extractMutationSequences(filteredUCSFirst100SNP, 15)
extracted_sequences <- extracted_data$sequences

# Calculate GC content for extracted sequences
gc_content <- calculateGCContent(extracted_sequences)
print(gc_content)
}

}
\references{
Pagès H, Aboyoun P, Gentleman R, DebRoy S (2024). \emph{Biostrings: Efficient manipulation of biological
strings}. doi:10.18129/B9.bioc.Biostrings \url{https://doi.org/10.18129/B9.bioc.Biostrings}, R package version
2.72.1, \url{https://bioconductor.org/packages/Biostrings}.
}
