% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractMutationData.R
\name{extractMutationData}
\alias{extractMutationData}
\title{Extract Specific Columns from Mutation Data}
\usage{
extractMutationData(mutation_data, selected_columns)
}
\arguments{
\item{mutation_data}{A data frame containing mutation data.}

\item{selected_columns}{A character vector specifying the columns to be extracted from the
mutation_data data frame.}
}
\value{
Returns a data frame containing only the columns specified by the user.
}
\description{
This function extracts specified columns from a mutation dataset. It checks whether the input
is a data frame and whether the specified columns exist in the data frame. It is assumed that
the dataset has been pre-loaded and is provided as an argument to the function.
}
\examples{
\dontrun{
library(RTCGA.mutations)
data(UCS.mutations)
extractMutationData(UCS.mutations, c("Chromosome", "Start_position", "End_position"))
}

}
