% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{filteredUCSFirst100SNP}
\alias{filteredUCSFirst100SNP}
\title{The First 100 Single Nucleotide Polymorphism Mutations In The UCS.mutations Dataset}
\format{
A data frame with columns:
\describe{
\item{Chromosome}{The chromosome on which the mutation is located.}
\item{Start_position}{The starting position of the mutation on the chromosome.}
\item{Reference_Allele}{The base(s) found on the reference genome at the mutation position.}
\item{Tumor_Seq_Allele2}{The observed allele sequence in the tumor sample at the mutation position.}
\item{Tumor_Sample_Barcode}{A unique identifier for the tumor sample in which the mutation was detected.}
\item{SNP_mutation}{The last three characters of the \code{Genome_Change} column, representing the single nucleotide polymorphism (SNP) mutation type (e.g., "G>T").}
\item{Variant_Type}{The type of mutation, such as single nucleotide polymorphism (SNP), deletion, or insertion.}
}
}
\source{
The Cancer Genome Atlas (TCGA) Program.
}
\usage{
filteredUCSFirst100SNP
}
\description{
Uterine Carcinosarcoma (UCS) Mutations datasets from The Cancer Genome Atlas Project 2015-11-01 snapshot.
}
\examples{
\dontrun{
 filteredUCSFirst100SNP
}
}
\references{
#' Kosinski M (2024). \emph{RTCGA.mutations: Mutations datasets from The Cancer Genome Atlas Project}.
doi:10.18129/B9.bioc.RTCGA.mutations \url{https://doi.org/10.18129/B9.bioc.RTCGA.mutations}, R package version
20151101.34.0, \url{https://bioconductor.org/packages/RTCGA.mutations}
}
\keyword{datasets}
