% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runCancerMutAnalyzer.R
\name{runCancerMutAnalyzer}
\alias{runCancerMutAnalyzer}
\title{Launch Shiny App For Package CancerMutAnalyzer}
\usage{
runCancerMutAnalyzer()
}
\value{
No return value but open up a shiny page.
}
\description{
A function that launches the Shiny app for CancerMutAnalyzer
This app will allow users to create visualisations of Mutation Frequency
based on different clinical attributes.
The code has been placed in \code{./inst/shiny-scripts}.
}
\examples{
\dontrun{
runCancerMutAnalyzer()
}

}
\references{
Grolemund, G. (2015). Learn Shiny - Video Tutorials. \href{https://shiny.rstudio.com/tutorial/}{Link}
}
\author{
Keren Zhang, \email{keren.zhang@utoronto.ca}
}
