% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{UCS.mutations}
\alias{UCS.mutations}
\title{UCS.mutations Dataset From the RTCGA.mutations Package}
\format{
A data frame with multiple rows and the following columns:
\describe{
\item{Hugo_Symbol}{Gene symbol associated with the mutation.}
\item{Entrez_Gene_Id}{Entrez gene ID for the mutated gene.}
\item{Center}{Research center that generated the data.}
\item{NCBI_Build}{NCBI genome build version used for alignment (e.g., hg19).}
\item{Chromosome}{The chromosome on which the mutation is located (e.g., "chr1", "chr2").}
\item{Start_position}{The starting genomic position of the mutation.}
\item{End_position}{The ending genomic position of the mutation.}
\item{Strand}{Strand on which the mutation occurs, either "+" or "-".}
\item{Variant_Classification}{Classification of the mutation (e.g., missense, nonsense).}
\item{Variant_Type}{The type of mutation, such as SNP, insertion, or deletion.}
\item{Reference_Allele}{The reference allele sequence at the mutation site.}
\item{Tumor_Seq_Allele1}{The sequence of the first tumor allele at the mutation site.}
\item{Tumor_Seq_Allele2}{The sequence of the second tumor allele at the mutation site.}
\item{dbSNP_RS}{Reference SNP ID from dbSNP database.}
\item{Tumor_Sample_Barcode}{Unique identifier for the tumor sample containing the mutation.}
}
}
\source{
The Cancer Genome Atlas (TCGA) Program.
}
\usage{
UCS.mutations
}
\description{
Mutations datasets from The Cancer Genome Atlas Project 2015-11-01 snapshot.
}
\examples{
\dontrun{
 UCS.mutations
}
}
\keyword{datasets}
