% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{UCS.mutations_snp_first_100}
\alias{UCS.mutations_snp_first_100}
\title{The First 100 Single Nucleotide Polymorphism Mutations In The UCS.mutations Dataset}
\format{
A data frame with columns:
\describe{
\item{Chromosome}{The chromosome on which the mutation is located.}
\item{Start_position}{The starting position of the mutation on the chromosome.}
\item{Reference_Allele}{The base(s) found on the reference genome at the mutation position.}
\item{Tumor_Seq_Allele2}{The observed allele sequence in the tumor sample at the mutation position.}
\item{Tumor_Sample_Barcode}{A unique identifier for the tumor sample in which the mutation was detected.}
\item{Variant_Type}{The type of mutation, such as single nucleotide polymorphism (SNP), deletion, or insertion.}
}
}
\source{
The Cancer Genome Atlas (TCGA) Program.
}
\usage{
UCS.mutations_snp_first_100
}
\description{
Uterine Carcinosarcoma (UCS) Mutations datasets from The Cancer Genome Atlas Project 2015-11-01 snapshot.
}
\examples{
\dontrun{
 UCS.mutations_snp_first_100
}
}
\keyword{datasets}
