% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateGCContent.R
\name{calculateGCContent}
\alias{calculateGCContent}
\title{Calculate GC Content of Extracted Sequences}
\usage{
calculateGCContent(sequences)
}
\arguments{
\item{sequences}{A character vector of DNA sequences, typically obtained from the \code{extractMutationSequences} function.}
}
\value{
A numeric vector with GC content percentages for each sequence in the input.
}
\description{
This function calculates the GC content (the percentage of G and C bases) for each sequence around a mutation site.
}
\details{
This function assumes that the input sequences are in a character vector format, with each element representing a DNA sequence.
The GC content is calculated as the percentage of 'G' and 'C' bases in each sequence, relative to the total length of the sequence.
}
\examples{
\dontrun{
# Example 1
# Calculate GC content for a predefined set of sequences
sequences <- c("ATGCGT", "CGCGGC", "ATATAT")
gc_content <- calculateGCContent(sequences)
print(gc_content)

# Example 2
# Extract sequences around mutation sites and calculate their GC content
library(BSgenome.Hsapiens.UCSC.hg19)
library(GenomicRanges)


# Extract sequences
extracted_data <- extractMutationSequences(UCS.mutations_snp_first_100, 15)
extracted_sequences <- extracted_data$sequences

# Calculate GC content for extracted sequences
gc_content <- calculateGCContent(extracted_sequences)
print(gc_content)
}

}
