% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processMutationData.R
\name{extractMutationData}
\alias{extractMutationData}
\title{Extract Specific Columns from Mutation Data}
\usage{
extractMutationData(
  mutation_data,
  selected_columns = c("Chromosome", "Start_position", "End_position", "Variant_Type",
    "Reference_Allele", "Tumor_Seq_Allele1", "Genome_Change", "Tumor_Sample_Barcode")
)
}
\arguments{
\item{mutation_data}{Cancer mutation data in Mutation Annotation Format (MAF) or data frame format.}

\item{selected_columns}{A character vector specifying the columns to be extracted from the
mutation_data data frame. Default columns are:
\itemize{
\item \code{"Chromosome"}: The chromosome number on which the mutation occurs.
\item \code{"Start_position"}: The starting position of the mutation on the chromosome.
\item \code{"End_position"}: The ending position of the mutation on the chromosome.
\item \code{"Variant_Type"}: The type of mutation, such as single nucleotide polymorphisms (SNP), deletion, or insertion.
\item \code{"Reference_Allele"}: The base(s) found on the reference allele at the position of the mutation.
\item \code{"Tumor_Seq_Allele1"}: The base(s) found in the tumor sample at the mutation position.
\item \code{"Genome_Change"}: A character vector specifying the genomic change at the mutation site, typically in the format
"g.chr\if{html}{\out{<chromosome>}}:\if{html}{\out{<position>}}<reference_base>>><mutated_base>". For example, "g.chr5:61733126G>T" indicates
a mutation on chromosome 5 at position 61733126, where the reference base "G" has been mutated to "T".
\item \code{"Tumor_Sample_Barcode"}: A unique identifier for the tumor sample in which the mutation was detected.
}}
}
\value{
Returns a data frame containing only the columns specified by the user.
}
\description{
This function extracts specified columns from a mutation data set. It has default settings
for commonly used columns but allows for user customization. It checks whether the input
is a data frame and whether the specified columns exist in the data frame.
}
\examples{
# Example 1

\dontrun{
library(RTCGA.mutations)
data(UCS.mutations)
extractMutationData(UCS.mutations, c("Chromosome", "Start_position", "End_position"))
}

# Example 2

\dontrun{
library(RTCGA.mutations)
data(UCS.mutations)
extractMutationData(UCS.mutations)
}

}
\references{
Kosinski M (2024). \emph{RTCGA.mutations: Mutations datasets from The Cancer Genome Atlas Project}.
doi:10.18129/B9.bioc.RTCGA.mutations \url{https://doi.org/10.18129/B9.bioc.RTCGA.mutations}, R package version
20151101.34.0, \url{https://bioconductor.org/packages/RTCGA.mutations}.

R Core Team (2024). \emph{R: A Language and Environment for Statistical Computing}. R Foundation for Statistical
Computing, Vienna, Austria. \url{https://www.R-project.org/}.

Wickham H, François R, Henry L, Müller K, Vaughan D (2023). \emph{dplyr: A Grammar of Data Manipulation}. R
package version 1.1.4, \url{https://CRAN.R-project.org/package=dplyr}.
}
