% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processMutationData.R
\name{extractMutationData}
\alias{extractMutationData}
\title{Extract Specific Columns from Mutation Data}
\usage{
extractMutationData(
  mutation_data,
  selected_columns = c("Chromosome", "Start_position", "End_position", "Variant_Type",
    "Reference_Allele", "Tumor_Seq_Allele1", "Genome_Change", "Tumor_Sample_Barcode")
)
}
\arguments{
\item{mutation_data}{A data frame containing detailed cancer mutation information. It should be structured with columns that can include various mutation attributes. This data frame is expected to be sourced from genomic studies databases like TCGA, pre-loaded into the R environment using the \code{data()} function.}

\item{selected_columns}{A character vector specifying the columns to be extracted. The default selection includes key mutation attributes commonly analyzed in genomic research:
\itemize{
\item \code{"Chromosome"}: The chromosome number where the mutation is located.
\item \code{"Start_position"}: The genomic start position of the mutation.
\item \code{"End_position"}: The genomic end position of the mutation.
\item \code{"Variant_Type"}: The type of genetic variation (e.g., SNP, deletion).
\item \code{"Reference_Allele"}: The reference allele at the mutation position.
\item \code{"Tumor_Seq_Allele1"}: The allele in the tumor sample.
\item \code{"Genome_Change"}: A description of the genomic alteration.
\item \code{"Tumor_Sample_Barcode"}: Identifier for the tumor sample.
}}
}
\value{
A data frame containing only the selected columns, facilitating focused analysis on specific mutation attributes.
}
\description{
This function extracts specified columns from a dataset containing genetic mutation data. It's tailored to handle data in Mutation Annotation Format (MAF) or similar structured data frames, which are typical outputs of genomic studies such as those from TCGA. This function simplifies the process of selecting specific mutation-related attributes for further analysis.
}
\examples{
\dontrun{
library(RTCGA.mutations)
data(UCS.mutations) # Load the UCS.mutations dataset from the RTCGA.mutations package. This dataset contains mutation data for uterine carcinosarcoma and is used here to demonstrate data extraction.
extractMutationData(UCS.mutations, c("Chromosome", "Start_position", "End_position"))
# This example extracts the Chromosome, Start_position, and End_position columns from the UCS.mutations dataset to analyze genomic intervals.

extractMutationData(UCS.mutations)
# In this example, the function uses default parameters to extract a broader set of mutation characteristics for initial explorations or broader analyses.
}

}
\references{
Kosinski M (2024). \emph{RTCGA.mutations: Mutations datasets from The Cancer Genome Atlas Project}.
doi:10.18129/B9.bioc.RTCGA.mutations \url{https://doi.org/10.18129/B9.bioc.RTCGA.mutations}, R package version
20151101.34.0, \url{https://bioconductor.org/packages/RTCGA.mutations}.

R Core Team (2024). \emph{R: A Language and Environment for Statistical Computing}. R Foundation for Statistical
Computing, Vienna, Austria. \url{https://www.R-project.org/}.

Wickham H, François R, Henry L, Müller K, Vaughan D (2023). \emph{dplyr: A Grammar of Data Manipulation}. R
package version 1.1.4, \url{https://CRAN.R-project.org/package=dplyr}.
}
