% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processMutationData.R
\name{filterMutations}
\alias{filterMutations}
\title{Filter Mutation Data by Specified Conditions}
\usage{
filterMutations(data, conditions)
}
\arguments{
\item{data}{A data frame containing mutation data, such as \code{UCS.mutations}.}

\item{conditions}{A list of conditions to filter by, where each element is named with a column name
and contains the value or range to filter on. Example: \code{list(Chromosome = "1", Variant_Type = "SNP")}.}
}
\value{
A filtered data frame that matches the specified conditions.
}
\description{
This function filters mutation data based on user-specified conditions for any columns in the dataset.
}
\examples{
\dontrun{

# Example 1
# Filter by chromosome and variant type
filtered_data <- filterMutations(UCS.mutations, conditions = list(Chromosome = "1", Variant_Type = "SNP"))
filtered_data

# Example 2
# Filter by a specific gene symbol
filtered_data <- filterMutations(UCS.mutations, conditions = list(Hugo_Symbol = "TP53"))
}

}
