% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizeMutationFrequency.R
\name{visualizeMutationFrequencyBar}
\alias{visualizeMutationFrequencyBar}
\title{Visualize Mutation Frequency with a Bar Plot}
\usage{
visualizeMutationFrequencyBar(data, group_by_column)
}
\arguments{
\item{data}{A data frame containing mutation data, which must include the specified grouping column.}

\item{group_by_column}{A character string specifying the column to group by for mutation frequency counts.}
}
\value{
A ggplot2 object showing a bar plot with the specified grouping column on the x-axis and
the mutation frequency on the y-axis.
}
\description{
This function generates a bar plot showing the frequency of mutations based on a specified column.
It is ideal for visualizing the distribution of mutations by a single category, such as chromosome or allele type.
}
\examples{
\dontrun{

# Example 1
# Visualize mutation frequency by chromosome
visualizeMutationFrequencyBar(UCS.mutations, group_by_column = "Chromosome")

Example 2
visualizeMutationFrequencyBar(filteredUCSFirst100SNP, group_by_column = "SNP_Mutation")
}

}
