% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualizeMutationFrequency.R
\name{visualizeMutationFrequencyHeatmap}
\alias{visualizeMutationFrequencyHeatmap}
\title{Visualize Mutation Frequency with a Heatmap}
\usage{
visualizeMutationFrequencyHeatmap(data, group_by_columns)
}
\arguments{
\item{data}{A data frame containing mutation data, which must include the two specified grouping columns.}

\item{group_by_columns}{A character vector of exactly two column names to group by for mutation frequency counts.
These columns must contain only nucleotide bases (e.g., "A", "T", "C", "G").}
}
\value{
A ggplot2 object showing a heatmap with one column on the x-axis and the other on the y-axis.
The color intensity represents the frequency of mutations.
}
\description{
This function creates a heatmap to visualize mutation frequency based on two specified columns,
such as \code{Reference_Allele} and \code{Tumor_Seq_Allele2}. Rows with values other than single-letter bases
in either column are excluded.
}
\examples{
\dontrun{
# Visualize mutation frequency by Reference and Tumor alleles
visualizeMutationFrequencyHeatmap(UCS.mutations, group_by_columns = c("Reference_Allele", "Tumor_Seq_Allele2"))
}

}
